+++ source/isaaclab/isaaclab/scene/interactive_scene_cfg.py
@@ -50,7 +50,8 @@ class InteractiveSceneCfg:
 
             # sensor - ray caster attached to the base of robot 1 that scans the ground
             height_scanner = RayCasterCfg(
-                prim_path="{ENV_REGEX_NS}/Robot_1/base",
+                prim_path="{ENV_REGEX_NS}/Robot_1/base/raycaster",
+                spawn=sim_utils.RayCasterXformCfg(),
                 offset=RayCasterCfg.OffsetCfg(pos=(0.0, 0.0, 20.0)),
                 ray_alignment="yaw",
                 pattern_cfg=GridPatternCfg(resolution=0.1, size=[1.6, 1.0]),