+++ source/isaaclab/isaaclab/sensors/ray_caster/ray_cast_utils.py
@@ -10,13 +10,11 @@ from __future__ import annotations
 import torch
 import warp as wp
 
-import omni.physics.tensors.impl.api as physx
-
 from isaaclab.sim.views import BaseXformPrimView
 
 
 def obtain_world_pose_from_view(
-    physx_view: BaseXformPrimView | physx.ArticulationView | physx.RigidBodyView,
+    physx_view: BaseXformPrimView,
     env_ids: torch.Tensor,
     clone: bool = False,
 ) -> tuple[torch.Tensor, torch.Tensor]:
@@ -30,9 +28,6 @@ def obtain_world_pose_from_view(
     Returns:
         A tuple containing the world positions and orientations of the prims.
         Orientation is in (x, y, z, w) format.
-
-    Raises:
-        NotImplementedError: If the prim view is not of the supported type.
     """
     indices = wp.from_torch(env_ids.to(dtype=torch.int32), dtype=wp.int32) if env_ids is not None else None
     pos_wp, quat_wp = physx_view.get_world_poses(indices)