+++ source/isaaclab/test/markers/check_markers_visibility.py
@@ -67,7 +67,8 @@ class SensorsSceneCfg(InteractiveSceneCfg):
 
     # sensors
     height_scanner = RayCasterCfg(
-        prim_path="{ENV_REGEX_NS}/Robot/base",
+        prim_path="{ENV_REGEX_NS}/Robot/base/raycaster",
+        spawn=sim_utils.RayCasterXformCfg(),
         update_period=0.02,
         offset=RayCasterCfg.OffsetCfg(pos=(0.0, 0.0, 20.0)),
         ray_alignment="yaw",